Fix a retain cycle of STHTTPRequest objects in STTwitterOAuth.
Under ARC (and not under MRR), variables declared __block are retained
when captured in a copied block. The solution is to another __block
__weak declared variable and use that within the blocks copied by the
STHTTPRequest.

This issue is particularly bad if you're posting images, as I was,
because the image data is retained by the STHTTPRequest. So every
image post you leak the whole image data.