Use the new value of `e` in the final errorBlock.
There is an earlier call to errorBlock that consumers `error`, but is gated
by an `if (error)` check. `error` doesn't change in the meantime, so the second
call doesn't make sense. On top of that, the second value of `e` was going
unused. It seems like the final errorBlock intends to pass `e`, as the new value
of `e` is packaging the responseString, which is the most likely source of an
explanation for why the response triggered the outer errorBlock instead of its
completionBlock.